backend/requirements.txt
# FastAPI and server
fastapi
uvicorn
pydantic
firebase-admin

# File uploads
python-multipart

# AI Model inference
torch
torchvision
Pillow
numpy

# Database & Storage
minio
opencv-python-headless
librosa
transformers
groq
python-dotenv
